+++ src/webbin/models.py
@@ -1,14 +1,25 @@
 from __future__ import annotations
 
-from datetime import datetime
-from typing import Any
+import sys
 
 from . import db
 
-Model: Any = db.Model
+Model = db.Model
 
 
-class Paste(Model):
+if sys.version_info >= (3, 11):
+    from datetime import UTC, datetime
+
+    def get_utc_now() -> datetime:
+        return datetime.now(UTC)
+else:
+    from datetime import datetime
+
+    def get_utc_now() -> datetime:
+        return datetime.utcnow()
+
+
+class Paste(Model):  # type: ignore[valid-type,misc]
     __tablename__ = "pastes"
     id = db.Column(db.String(32), unique=True, index=True, primary_key=True)
     title = db.Column(db.LargeBinary(100))
@@ -20,7 +31,7 @@ class Paste(Model):
     created_at: datetime = db.Column(
         db.DateTime(),
         nullable=False,
-        default=datetime.utcnow,
+        default=get_utc_now,
     )
     expires_after = db.Column(db.Interval(), nullable=False)
     views = db.Column(db.Integer(), nullable=False, default=0)
