+++ src/webbin/static/clipboard.js
